+++ src/main/kotlin/com/nowsprinting/intellij_mob/action/done/DoneTask.kt
@@ -11,6 +11,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.Task.Backgroundable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.nowsprinting.intellij_mob.MobBundle
 import com.nowsprinting.intellij_mob.config.MobProjectSettings
 import com.nowsprinting.intellij_mob.config.validateForDoneTask
@@ -52,7 +53,7 @@ class DoneTask(val settings: MobProjectSettings, val e: AnActionEvent, project: 
             return
         }
 
-        val (validRepository, reasonInvalidRepository) = repository.validateForDonePrecondition(settings)
+        val (validRepository, reasonInvalidRepository) = repository.validateForDone(settings)
         if (!validRepository) {
             val format = MobBundle.message("mob.done.error.precondition")
             val message = String.format(format, reasonInvalidRepository)
@@ -65,8 +66,8 @@ class DoneTask(val settings: MobProjectSettings, val e: AnActionEvent, project: 
         stopTimer()
 
         val hasUncommittedChanges = hasUncommittedChanges(repository)
-        val hasUnpushedCommit = hasChangesForDone(settings, repository)
-        if (!hasUncommittedChanges && !hasUnpushedCommit) {
+        val hasChangesForDone = hasChangesForDone(settings, repository)
+        if (!hasUncommittedChanges && !hasChangesForDone) {
             val format = MobBundle.message("mob.done.error.precondition")
             val message = String.format(format, MobBundle.message("mob.done.error.reason.nothing_changes_to_squash"))
             logger.warn(message)
@@ -123,6 +124,9 @@ class DoneTask(val settings: MobProjectSettings, val e: AnActionEvent, project: 
                 contents = notifyContents,
                 type = NotificationType.ERROR
             )
+        }
+        VirtualFileManager.getInstance().asyncRefresh {
+            logger.debug(MobBundle.message("mob.logging.refresh"))
         }
     }