+++ src/main/kotlin/com/nowsprinting/intellij_mob/action/next/NextTask.kt
@@ -9,6 +9,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.Task.Backgroundable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.nowsprinting.intellij_mob.MobBundle
 import com.nowsprinting.intellij_mob.config.MobProjectSettings
 import com.nowsprinting.intellij_mob.config.validateForNextTask
@@ -50,7 +51,7 @@ class NextTask(val settings: MobProjectSettings, project: Project, title: String
             return
         }
 
-        val (validRepository, reasonInvalidRepository) = repository.validateForNextPrecondition(settings)
+        val (validRepository, reasonInvalidRepository) = repository.validateForNext(settings)
         if (!validRepository) {
             val format = MobBundle.message("mob.next.error.precondition")
             val message = String.format(format, reasonInvalidRepository)
@@ -134,6 +135,9 @@ class NextTask(val settings: MobProjectSettings, project: Project, title: String
                 contents = notifyContents,
                 type = NotificationType.ERROR
             )
+        }
+        VirtualFileManager.getInstance().asyncRefresh {
+            logger.debug(MobBundle.message("mob.logging.refresh"))
         }
     }