+++ src/main/kotlin/com/nowsprinting/intellij_mob/action/reset/ResetTask.kt
@@ -9,6 +9,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.Task.Backgroundable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.nowsprinting.intellij_mob.MobBundle
 import com.nowsprinting.intellij_mob.config.MobProjectSettings
 import com.nowsprinting.intellij_mob.config.validateForResetTask
@@ -48,7 +49,7 @@ class ResetTask(val settings: MobProjectSettings, project: Project, title: Strin
             return
         }
 
-        val (validRepository, reasonInvalidRepository) = repository.validateForResetPrecondition(settings)
+        val (validRepository, reasonInvalidRepository) = repository.validateForReset(settings)
         if (!validRepository) {
             val format = MobBundle.message("mob.reset.error.precondition")
             val message = String.format(format, reasonInvalidRepository)
@@ -105,6 +106,9 @@ class ResetTask(val settings: MobProjectSettings, project: Project, title: Strin
                 contents = notifyContents,
                 type = NotificationType.ERROR
             )
+        }
+        VirtualFileManager.getInstance().asyncRefresh {
+            logger.debug(MobBundle.message("mob.logging.refresh"))
         }
     }