+++ src/main/kotlin/com/nowsprinting/intellij_mob/action/start/StartTask.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.Task.Backgroundable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.nowsprinting.intellij_mob.MobBundle
 import com.nowsprinting.intellij_mob.action.share.ShareAction
 import com.nowsprinting.intellij_mob.config.MobProjectSettings
@@ -53,7 +54,7 @@ class StartTask(val settings: MobProjectSettings, val e: AnActionEvent, project:
             return
         }
 
-        val (validRepository, reasonInvalidRepository) = repository.validateForStartPrecondition(settings)
+        val (validRepository, reasonInvalidRepository) = repository.validateForStart(settings)
         if (!validRepository) {
             val format = MobBundle.message("mob.start.error.precondition")
             val message = String.format(format, reasonInvalidRepository)
@@ -120,6 +121,9 @@ class StartTask(val settings: MobProjectSettings, val e: AnActionEvent, project:
                 contents = notifyContents,
                 type = NotificationType.ERROR
             )
+        }
+        VirtualFileManager.getInstance().asyncRefresh {
+            logger.debug(MobBundle.message("mob.logging.refresh"))
         }
     }