+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
-- Add shortcut for done action (ALT+M, D)
+
+## 1.0-beta2
+- Add shortcut for show done dialog (ALT+M, D)
+- Fix bug: refresh project files after start/next/done/reset
+- Fix bug: notify error at done without changes
 
 ## 1.0-beta1
 - Display timer on statusbar